+++ packages/device_info_plus/lib/device_info_plus_tizen.dart
@@ -1,11 +1,9 @@
 // Copyright 2020 Samsung Electronics Co., Ltd. All rights reserved.
-// Copyright 2013 The Flutter Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
 import 'dart:async';
 
-import 'package:device_info_plus_platform_interface/device_info_plus_platform_interface.dart';
 import 'package:flutter/services.dart';
 
 /// Information derived from `system_info`.
@@ -19,6 +17,7 @@ class TizenDeviceInfo {
     required this.nativeApiVersion,
     required this.platformVersion,
     required this.webApiVersion,
+    required this.profile,
     required this.buildDate,
     required this.buildId,
     required this.buildString,
@@ -28,110 +27,101 @@ class TizenDeviceInfo {
     required this.buildRelease,
     required this.deviceType,
     required this.manufacturer,
+    required this.platformName,
+    required this.platformProcessor,
+    required this.tizenId,
   });
 
   /// http://tizen.org/system/model_name
-  ///
-  /// The value is an empty String if it is not available.
-  final String modelName;
+  final String? modelName;
 
   /// http://tizen.org/feature/platform.core.cpu.arch
-  ///
-  /// The value is an empty String if it is not available.
-  final String cpuArch;
+  final String? cpuArch;
 
   /// http://tizen.org/feature/platform.native.api.version
-  ///
-  /// The value is an empty String if it is not available.
-  final String nativeApiVersion;
+  final String? nativeApiVersion;
 
   /// http://tizen.org/feature/platform.version
-  ///
-  /// The value is an empty String if it is not available.
-  final String platformVersion;
+  final String? platformVersion;
 
   /// http://tizen.org/feature/platform.web.api.version
-  ///
-  /// The value is an empty String if it is not available.
-  final String webApiVersion;
+  final String? webApiVersion;
+
+  /// http://tizen.org/feature/profile
+  final String? profile;
 
   /// http://tizen.org/system/build.date
-  ///
-  /// The value is an empty String if it is not available.
-  final String buildDate;
+  final String? buildDate;
 
   /// http://tizen.org/system/build.id
-  ///
-  /// The value is an empty String if it is not available.
-  final String buildId;
+  final String? buildId;
 
   /// http://tizen.org/system/build.string
-  ///
-  /// The value is an empty String if it is not available.
-  final String buildString;
+  final String? buildString;
 
   /// http://tizen.org/system/build.time
-  ///
-  /// The value is an empty String if it is not available.
-  final String buildTime;
+  final String? buildTime;
 
   /// http://tizen.org/system/build.type
-  ///
-  /// The value is an empty String if it is not available.
-  final String buildType;
+  final String? buildType;
 
   /// http://tizen.org/system/build.variant
-  ///
-  /// The value is an empty String if it is not available.
-  final String buildVariant;
+  final String? buildVariant;
 
   /// http://tizen.org/system/build.release
-  ///
-  /// The value is an empty String if it is not available.
-  final String buildRelease;
+  final String? buildRelease;
 
   /// http://tizen.org/system/device_type
-  ///
-  /// The value is an empty String if it is not available.
-  final String deviceType;
+  final String? deviceType;
 
   /// http://tizen.org/system/manufacturer
-  ///
-  /// The value is an empty String if it is not available.
-  final String manufacturer;
+  final String? manufacturer;
+
+  /// http://tizen.org/system/platform.name
+  final String? platformName;
+
+  /// http://tizen.org/system/platform.processor
+  final String? platformProcessor;
+
+  /// http://tizen.org/system/tizenid
+  final String? tizenId;
 
   /// Deserializes from the message received from [_kChannel].
   static TizenDeviceInfo fromMap(Map<String, dynamic> map) {
     return TizenDeviceInfo(
-      modelName: map['modelName'] as String? ?? '',
-      cpuArch: map['cpuArch'] as String? ?? '',
-      nativeApiVersion: map['nativeApiVersion'] as String? ?? '',
-      platformVersion: map['platformVersion'] as String? ?? '',
-      webApiVersion: map['webApiVersion'] as String? ?? '',
-      buildDate: map['buildDate'] as String? ?? '',
-      buildId: map['buildId'] as String? ?? '',
-      buildString: map['buildString'] as String? ?? '',
-      buildTime: map['buildTime'] as String? ?? '',
-      buildType: map['buildType'] as String? ?? '',
-      buildVariant: map['buildVariant'] as String? ?? '',
-      buildRelease: map['buildRelease'] as String? ?? '',
-      deviceType: map['deviceType'] as String? ?? '',
-      manufacturer: map['manufacturer'] as String? ?? '',
+      modelName: map['modelName'],
+      cpuArch: map['cpuArch'],
+      nativeApiVersion: map['nativeApiVersion'],
+      platformVersion: map['platformVersion'],
+      webApiVersion: map['webApiVersion'],
+      profile: map['profile'],
+      buildDate: map['buildDate'],
+      buildId: map['buildId'],
+      buildString: map['buildString'],
+      buildTime: map['buildTime'],
+      buildType: map['buildType'],
+      buildVariant: map['buildVariant'],
+      buildRelease: map['buildRelease'],
+      deviceType: map['deviceType'],
+      manufacturer: map['manufacturer'],
+      platformName: map['platformName'],
+      platformProcessor: map['platformProcessor'],
+      tizenId: map['tizenId'],
     );
   }
 }
 
-/// An implementation of [DeviceInfoPlatform] that uses method channels.
-class MethodChannelDeviceInfoTizen extends DeviceInfoPlatform {
+class _MethodChannelDeviceInfo {
   /// The method channel used to interact with the native platform.
   MethodChannel channel =
       const MethodChannel('dev.fluttercommunity.plus/device_info');
 
-  /// Method channel for Tizen devices
+  /// Method channel for Tizen devices.
   Future<TizenDeviceInfo> tizenInfo() async {
-    return TizenDeviceInfo.fromMap((await channel
-            .invokeMapMethod<String, dynamic>('getTizenDeviceInfo')) ??
-        <String, dynamic>{});
+    return TizenDeviceInfo.fromMap(
+      (await channel.invokeMethod('getTizenDeviceInfo'))
+          .cast<String, dynamic>(),
+    );
   }
 }
 
@@ -141,8 +131,7 @@ class DeviceInfoPluginTizen {
   /// repeatedly or in performance-sensitive blocks.
   DeviceInfoPluginTizen();
 
-  static final MethodChannelDeviceInfoTizen _tizenInstance =
-      MethodChannelDeviceInfoTizen();
+  static final _MethodChannelDeviceInfo _platform = _MethodChannelDeviceInfo();
 
   /// This information does not change from call to call. Cache it.
   TizenDeviceInfo? _cachedTizenDeviceInfo;
@@ -151,5 +140,5 @@ class DeviceInfoPluginTizen {
   ///
   /// See: https://docs.tizen.org/application/native/guides/device/system
   Future<TizenDeviceInfo> get tizenInfo async =>
-      _cachedTizenDeviceInfo ??= await _tizenInstance.tizenInfo();
+      _cachedTizenDeviceInfo ??= await _platform.tizenInfo();
 }