+++ packages/device_info_plus/CHANGELOG.md
@@ -1,3 +1,11 @@
 ## 1.0.0
 
-* Initial release
+* Initial release.
+
+## 1.1.0
+
+* Remove dependency on `device_info_plus_platform_interface`.
+* Make values of `TizenDeviceInfo` nullable.
+* Add support for `profile`, `platformName`, `platformProcessor`, and `tizenId`.
+* Update the example app.
+* Minor cleanups.